+++ changelog.md
@@ -1,2 +1,5 @@
+# Upcoming
+* Added `get_random_rune` sub command
+
 # 1.0.0-beta
 * First Release